[refactoring] Use std::uint32_t instead of std::size_t for number of vertices/indices.

+++ include/inexor/vulkan-renderer/mesh_buffer.hpp
@@ -51,12 +51,12 @@ public:
 
     /// @brief Creates a new vertex buffer and an associated index buffer.
     MeshBuffer(const VkDevice device, VkQueue data_transfer_queue, const std::uint32_t data_transfer_queue_family_index, const VmaAllocator vma_allocator, std::string name,
-               const VkDeviceSize size_of_vertex_structure, const std::size_t number_of_vertices, void *vertices, const VkDeviceSize size_of_index_structure,
-               const std::size_t number_of_indices, void *indices);
+               const VkDeviceSize size_of_vertex_structure, const std::uint32_t number_of_vertices, void *vertices, const VkDeviceSize size_of_index_structure,
+               const std::uint32_t number_of_indices, void *indices);
 
     /// @brief Creates a vertex buffer without index buffer.
     MeshBuffer(const VkDevice device, VkQueue data_transfer_queue, const std::uint32_t data_transfer_queue_family_index, const VmaAllocator vma_allocator, std::string name,
-               const VkDeviceSize size_of_vertex_structure, const std::size_t number_of_vertices, void *vertices);
+               const VkDeviceSize size_of_vertex_structure, const std::uint32_t number_of_vertices, void *vertices);
 
     VkBuffer get_vertex_buffer() const {
         return vertex_buffer.get_buffer();
